Adding Documentation and handling player invalid input

+++ AutoBattleC++/AutoBattleC++.cpp
@@ -6,8 +6,8 @@
 
 int main()
 {
-	/*Makes the Player and Enemy appear at a different random spot each playtrough*/
-	srand(time(NULL));
+	/*Makes the Player and Enemy appear at a different random spot each playtrough */
+	srand((unsigned int)time(NULL));
 
     BattleField* battleField = new(BattleField);
 	battleField->Setup();
